+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/DebuggerSession.java
@@ -65,6 +65,7 @@ import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.RootCallTarget;
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleContext;
 import com.oracle.truffle.api.debug.Breakpoint.BreakpointConditionFailure;
@@ -80,6 +81,7 @@ import com.oracle.truffle.api.instrumentation.EventBinding;
 import com.oracle.truffle.api.instrumentation.EventContext;
 import com.oracle.truffle.api.instrumentation.ExecutionEventNode;
 import com.oracle.truffle.api.instrumentation.ExecutionEventNodeFactory;
+import com.oracle.truffle.api.instrumentation.InstrumentableNode;
 import com.oracle.truffle.api.instrumentation.ProbeNode;
 import com.oracle.truffle.api.instrumentation.SourceSectionFilter;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
@@ -381,6 +383,89 @@ public final class DebuggerSession implements Closeable {
 
         suspendNext = true;
         updateStepping();
+    }
+
+    /**
+     * Suspend immediately at the current location of the current execution thread. A {@link Node}
+     * argument can be provided as an exact current location, if known. This method can be called on
+     * the guest execution thread only.
+     * <p>
+     * This method calls {@link SuspendedCallback#onSuspend(SuspendedEvent)} synchronously, with
+     * {@link SuspendedEvent} created at the actual location of the current thread. This method can
+     * not be called from an existing {@link SuspendedCallback#onSuspend(SuspendedEvent) callback}.
+     *
+     * @param node the top Node of the execution, or <code>null</code>
+     * @return <code>true</code> when there is a guest code execution on the current thread and
+     *         {@link SuspendedCallback} was called, <code>false</code> otherwise.
+     * @throws IllegalStateException when the current thread is suspended already
+     * @throws IllegalArgumentException when a node with no {@link RootNode} is provided, or it's
+     *             root node does not match the current execution root, or the node does not match
+     *             the current call node, if known.
+     * @since 20.3
+     */
+    public boolean suspendHere(Node node) {
+        SuspendedEvent event = currentSuspendedEventMap.get(Thread.currentThread());
+        if (event != null) {
+            throw new IllegalStateException("Suspended already");
+        }
+        RootNode nodeRoot = null;
+        if (node != null) {
+            nodeRoot = node.getRootNode();
+            if (nodeRoot == null) {
+                throw new IllegalArgumentException(String.format("The node %s does not have a root.", node));
+            }
+        }
+        FrameInstance frameInstance = Truffle.getRuntime().getCurrentFrame();
+        if (frameInstance == null) {
+            return false;
+        }
+        RootNode root = ((RootCallTarget) frameInstance.getCallTarget()).getRootNode();
+        if (!includeInternal) {
+            if (root.isInternal()) {
+                // In an internal code, we need to iterate stack to find non-internal location
+                frameInstance = Truffle.getRuntime().iterateFrames(new FrameInstanceVisitor<FrameInstance>() {
+                    int startDepth = 1;
+
+                    @Override
+                    public FrameInstance visitFrame(FrameInstance fi) {
+                        if (startDepth-- > 0) {
+                            return null;
+                        }
+                        if (!((RootCallTarget) fi.getCallTarget()).getRootNode().isInternal()) {
+                            return fi;
+                        }
+                        return null;
+                    }
+                });
+                if (frameInstance == null) {
+                    return false;
+                } else {
+                    root = ((RootCallTarget) frameInstance.getCallTarget()).getRootNode();
+                }
+            }
+        }
+        if (nodeRoot != null && nodeRoot != root) {
+            throw new IllegalArgumentException(String.format("The node %s belongs to a root %s, which is different from the current root %s.", node, nodeRoot, root));
+        }
+        Node callNode = frameInstance.getCallNode();
+        if (callNode == null) {
+            callNode = node;
+            if (callNode == null) {
+                // We have no idea where in the function we are.
+                callNode = root;
+            }
+        }
+        if (node != null && node != callNode) {
+            throw new IllegalArgumentException(String.format("The node %s does not match the current known call node %s.", node, callNode));
+        }
+        Node icallNode = InstrumentableNode.findInstrumentableParent(callNode);
+        if (icallNode != null) {
+            callNode = icallNode;
+        }
+        MaterializedFrame frame = frameInstance.getFrame(FrameAccess.READ_WRITE).materialize();
+        SuspendedContext context = SuspendedContext.create(callNode, null);
+        doSuspend(context, SuspendAnchor.BEFORE, frame, null);
+        return true;
     }
 
     /**
@@ -997,7 +1082,7 @@ public final class DebuggerSession implements Closeable {
         // Fake the caller context
         Caller caller = findCurrentCaller(this, includeInternal);
         SuspendedContext context = SuspendedContext.create(caller.node, ((SteppingStrategy.Unwind) s).unwind);
-        doSuspend(context, SuspendAnchor.AFTER, caller.frame, insertableNode, null, null, null, Collections.emptyList(), Collections.emptyMap());
+        doSuspend(context, SuspendAnchor.AFTER, caller.frame, insertableNode);
     }
 
     static Caller findCurrentCaller(DebuggerSession session, boolean includeInternal) {
@@ -1137,9 +1222,15 @@ public final class DebuggerSession implements Closeable {
                 breakpointFailures.put(fb, failure.getConditionFailure());
             }
         }
+        if (breaks == null) {
+            breaks = Collections.emptyList();
+        }
+        if (breakpointFailures == null) {
+            breakpointFailures = Collections.emptyMap();
+        }
 
         boolean hitStepping = s.step(this, source.getContext(), suspendAnchor);
-        boolean hitBreakpoint = breaks != null && !breaks.isEmpty();
+        boolean hitBreakpoint = !breaks.isEmpty();
         Object newReturnValue = returnValue;
         if (hitStepping || hitBreakpoint) {
             s.consume();
@@ -1154,6 +1245,10 @@ public final class DebuggerSession implements Closeable {
             performKill(source.getContext().getInstrumentedNode());
         }
         return newReturnValue;
+    }
+
+    private Object doSuspend(SuspendedContext context, SuspendAnchor suspendAnchor, MaterializedFrame frame, InsertableNode insertableNode) {
+        return doSuspend(context, suspendAnchor, frame, insertableNode, null, null, null, Collections.emptyList(), Collections.emptyMap());
     }
 
     private Object doSuspend(SuspendedContext context, SuspendAnchor suspendAnchor, MaterializedFrame frame,